+++ tests/Fixtures/ConflictDifferentStructure/Controller.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\ConflictDifferentStructure;
+
+use Karewan\KnRoute\Attributes\Get;
+
+class Controller
+{
+	#[Get('/{everything:path}')]
+	public function everything(string $everything): void {}
+
+	#[Get('/users/{id:uint}')]
+	public function user(string $id): void {}
+}